requirements.txt
pytest~=8.3.5
beautifulsoup4~=4.13.3
soupsieve~=2.6
packaging~=24.2
h11~=0.14.0
pip~=25.0.1
attrs~=25.3.0
wheel~=0.45.1
PySocks~=1.7.1
filelock~=3.18.0
six~=1.17.0
setuptools~=70.3.0
rich~=13.9.4
pygments~=2.19.1
markdown-it-py~=3.0.0
trio~=0.29.0
outcome~=1.3.0.post0
sniffio~=1.3.1
exceptiongroup~=1.2.2
sortedcontainers~=2.4.0
idna~=3.10
PyYAML~=6.0.2
mdurl~=0.1.2
behave~=1.2.6
seleniumbase~=4.35.7
execnet~=2.1.1
pytest-xdist~=3.6.1
parse~=1.20.2
parse-type~=0.6.4
MarkupSafe~=3.0.2
pluggy~=1.5.0
iniconfig~=2.0.0
certifi~=2025.1.31
chardet~=5.2.0
tabcompleter~=1.4.0
urllib3~=1.26.20
charset-normalizer~=3.4.1
selenium~=4.29.0
trio-websocket~=0.12.2
websocket-client~=1.8.0
cssselect~=1.3.0
fasteners~=0.19
websockets~=15.0.1
jinja2~=3.1.6
platformdirs~=4.3.6
colorama~=0.4.6
requests~=2.32.3
mycdp~=1.1.1
pdbp~=1.6.1
parameterized~=0.9.0
wsproto~=1.2.0
sbvirtualdisplay~=1.4.0
python-dotenv~=1.0.1